+++ lib/ronin/support/binary/unhexdump/core_ext/file.rb
@@ -16,7 +16,7 @@
 # along with ronin-support.  If not, see <https://www.gnu.org/licenses/>.
 #
 
-require 'ronin/support/binary/hexdump/parser'
+require 'ronin/support/binary/unhexdump/parser'
 
 class File
 
@@ -28,7 +28,7 @@ class File
   #
   # @param [Hash{Symbol => Object}] kwargs
   #   Additional keyword arguments for
-  #   {Ronin::Support::Binary::Hexdump::Parser#initialize}.
+  #   {Ronin::Support::Binary::Unhexdump::Parser#initialize}.
   #
   # @option kwargs [:od, :hexdump] :format (:hexdump)
   #   The expected format of the hexdump. Must be either `:od` or
@@ -123,7 +123,7 @@ class File
   # @api public
   #
   def self.unhexdump(path,**kwargs)
-    parser = Ronin::Support::Binary::Hexdump::Parser.new(**kwargs)
+    parser = Ronin::Support::Binary::Unhexdump::Parser.new(**kwargs)
     parser.unhexdump(new(path))
   end